+++ src/GraphQL/Shopify/downloadListing2.graphql
@@ -1,0 +1,96 @@
+query GetProducts($numProducts: Int!, $cursor: String) {
+    products(first: $numProducts, after: $cursor) {
+        pageInfo {
+            hasNextPage
+            endCursor
+        }
+        nodes {
+            id
+            title
+            descriptionHtml
+            vendor
+            productType
+            createdAt
+            handle
+            updatedAt
+            publishedAt
+            templateSuffix
+            tags
+            status
+            seo {
+                title
+                description
+            }
+            variantsCount {
+                count
+                precision
+            }
+            variants(first: 200) {
+                pageInfo {
+                    hasNextPage
+                    endCursor
+                }
+                nodes {
+                    id
+                    title
+                    price
+                    position
+                    inventoryPolicy
+                    compareAtPrice
+                    selectedOptions {
+                        name
+                        value
+                    }
+                    createdAt
+                    updatedAt
+                    taxable
+                    barcode
+                    sku
+                    inventoryItem {
+                        id
+                    }
+                    inventoryQuantity
+                    image {
+                        id
+                        altText
+                        width
+                        height
+                        src
+                    }
+                }
+            }
+            options(first:2) {
+                id
+                name
+                position
+                values
+            }
+            mediaCount {
+                count
+                precision
+            }
+            media (first: 100) {
+                pageInfo {
+                    hasNextPage
+                    endCursor
+                }
+                nodes {
+                    id
+                    alt
+                    mediaContentType
+                    status
+                    preview {
+                        image {
+                            id
+                            altText
+                            width
+                            height
+                            url
+                        }
+                    }
+                }
+            }
+
+        }
+    }
+}
